Use a single JWT validator (#364)
