--- src/authorization/jwt/JwtIdentityProviderEnum.ts
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-/**
- * JWT Identity Providers. List all supported IDPs
- */
-export enum JwtIdentityProviderEnum {
-  MS_AAD = "https://sts.windows.net/72f988bf-86f1-41af-91ab-2d7cd011db47/",
-  Demo = "http://Demo-jwt-issuer",
-}
